Added ProjectTask.cs, ProjectTaskStatus.cs, ProjectTaskType.cs Worked on mapping, added IMapWith.cs, AssemblyMappingProfile.cs

+++ Jira.Domain/User.cs
@@ -1,12 +1,29 @@
 using Microsoft.AspNetCore.Identity;
 
 namespace Jira.Domain;
-
-public class User : IdentityUser<long>
+/// <summary>
+/// User 
+/// </summary>
+public class User : IdentityUser<int>
 {
+    /// <summary>
+    /// User's Family Name
+    /// </summary>
     public string LastName { get; set; }
+    /// <summary>
+    /// User's Name
+    /// </summary>
     public string FirstName { get; set; }
+    /// <summary>
+    /// User's Middle Name
+    /// </summary>
     public string MiddleName { get; set; }
+    /// <summary>
+    /// User's Email Address 
+    /// </summary>
     public string Email { get; set; }
+    /// <summary>
+    /// URL for user's profile avatar
+    /// </summary>
     public string AvatarUrl { get; set; }
 }
